+++ packages/amplify_authenticator/lib/src/enums/status_type.dart
@@ -15,7 +15,5 @@
 
 enum StatusType {
   info,
-  success,
-  warning,
   error,
 }