+++ packages/auth/amplify_auth_cognito_dart/lib/src/auth_plugin_impl.dart
@@ -15,6 +15,7 @@
 import 'dart:async';
 
 import 'package:amplify_auth_cognito_dart/amplify_auth_cognito_dart.dart';
+import 'package:amplify_auth_cognito_dart/src/credentials/device_metadata_repository.dart';
 import 'package:amplify_auth_cognito_dart/src/flows/constants.dart';
 import 'package:amplify_auth_cognito_dart/src/flows/helpers.dart';
 import 'package:amplify_auth_cognito_dart/src/flows/hosted_ui/initial_parameters_stub.dart'
@@ -115,6 +116,10 @@ class AmplifyAuthCognitoDart extends AuthPluginInterface
   /// The Cognito identity pool configuration.
   CognitoIdentityCredentialsProvider? get _identityPoolConfig =>
       _stateMachine.get();
+
+  /// The device metadata repository, used for handling device operations.
+  DeviceMetadataRepository get _deviceRepo =>
+      _stateMachine.getOrCreate(DeviceMetadataRepository.token);
 
   final StreamController<AuthHubEvent> _hubEventController =
       StreamController.broadcast();
@@ -750,7 +755,7 @@ class AmplifyAuthCognitoDart extends AuthPluginInterface
   }) async {
     final userPoolTokens = await getUserPoolTokens();
     final userId = userPoolTokens.idToken.userId;
-    final username = CognitoIdToken(userPoolTokens.idToken).username;
+    final username = userPoolTokens.username;
     return AuthUser(
       userId: userId,
       username: username,
@@ -759,21 +764,16 @@ class AmplifyAuthCognitoDart extends AuthPluginInterface
 
   @override
   Future<void> rememberDevice() async {
-    // Use credentials state machine since we need device info as well.
-    final result = await _stateMachine
-        .expect(CredentialStoreStateMachine.type)
-        .getCredentialsResult();
-    final deviceKey = result.data.deviceSecrets?.deviceKey;
+    final tokens = await getUserPoolTokens();
+    final username = tokens.username;
+    final deviceSecrets = await _deviceRepo.get(username);
+    final deviceKey = deviceSecrets?.deviceKey;
     if (deviceKey == null) {
       throw const DeviceNotTrackedException();
     }
-    final accessToken = result.data.userPoolTokens?.accessToken;
-    if (accessToken == null) {
-      throw const SignedOutException('No user is currently signed in');
-    }
     await _cognitoIdp.updateDeviceStatus(
       cognito.UpdateDeviceStatusRequest(
-        accessToken: accessToken.raw,
+        accessToken: tokens.accessToken.raw,
         deviceKey: deviceKey,
         deviceRememberedStatus: cognito.DeviceRememberedStatusType.remembered,
       ),
@@ -782,17 +782,17 @@ class AmplifyAuthCognitoDart extends AuthPluginInterface
 
   @override
   Future<void> forgetDevice([AuthDevice? device]) async {
-    // Use credentials state machine since we need device info as well.
-    final result = await _stateMachine
-        .expect(CredentialStoreStateMachine.type)
-        .getCredentialsResult();
-    final deviceKey = device?.id ?? result.data.deviceSecrets?.deviceKey;
+    final tokens = await getUserPoolTokens();
+    final username = tokens.username;
+    final deviceSecrets = await _deviceRepo.get(username);
+    final deviceKey = device?.id ?? deviceSecrets?.deviceKey;
     if (deviceKey == null) {
       throw const DeviceNotTrackedException();
     }
+    await _deviceRepo.remove(username);
     await _cognitoIdp.forgetDevice(
       cognito.ForgetDeviceRequest(
-        accessToken: result.data.userPoolTokens?.accessToken.raw,
+        accessToken: tokens.accessToken.raw,
         deviceKey: deviceKey,
       ),
     );
@@ -918,6 +918,7 @@ class AmplifyAuthCognitoDart extends AuthPluginInterface
     await _stateMachine.dispatch(
       const CredentialStoreEvent.clearCredentials(),
     );
+    await _deviceRepo.remove(tokens.username);
     _hubEventController
       ..add(AuthHubEvent.signedOut())
       ..add(AuthHubEvent.userDeleted());