+++ packages/auth/amplify_auth_cognito_dart/lib/src/flows/hosted_ui/hosted_ui_platform_io.dart
@@ -336,7 +336,7 @@ class HostedUiPlatformImpl extends HostedUiPlatform {
   /// Closes the open server, if any.
   @override
   Future<void> close() async {
-    await _localServer?.server.close(force: true);
+    await _localServer?.server.close();
     _localServer = null;
   }
 }