+++ packages/auth/amplify_auth_cognito_dart/lib/src/model/cognito_device_secrets.dart
@@ -36,6 +36,9 @@ abstract class CognitoDeviceSecrets
   /// The device key/ID.
   String get deviceKey;
 
+  /// The device password.
+  String get devicePassword;
+
   /// The [CognitoDeviceSecrets] serializer.
   static Serializer<CognitoDeviceSecrets> get serializer =>
       _$cognitoDeviceSecretsSerializer;