+++ packages/auth/amplify_auth_cognito_test/test/flows/srp/srp_password_verifier_worker_test.dart
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@OnPlatform({
-  'chrome': Skip('TODO(dnys1): Failing in CI'),
-})
-
 import 'dart:async';
 
 import 'package:amplify_auth_cognito_dart/src/flows/constants.dart';
@@ -31,7 +27,12 @@ import '../../common/mock_config.dart' hide username;
 import 'srp_helper_test.dart';
 
 void main() {
-  group('SrpPasswordVerifierWorker', () {
+  AWSLogger().logLevel = LogLevel.verbose;
+
+  // TODO(dnys1): This will skip tests on DDC where they're currently failing.
+  final skipTests = !zAssertsEnabled;
+
+  group('SrpPasswordVerifierWorker', skip: skipTests, () {
     test('success', () async {
       final worker = SrpPasswordVerifierWorker.create();
       addTearDown(worker.close);