+++ packages/auth/amplify_auth_cognito/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0-next.0+2
+
+- Fix package analysis issues
+
+## 1.0.0-next.0+1
+
+- Fix package analysis issues
+
 ## 1.0.0-next.0 (2022-08-02)
 
 Initial developer preview release for all platforms.